chat/serializers.py
from rest_framework import serializers
from .models import ChatMessage


class ChatSerializer(serializers.ModelSerializer):

    class Meta:
        model = ChatMessage
        fields = ['id', 'message', 'response', 'created_at']